Split unsafe code in smaller isolated pieces.
